requirements.txt
# requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
datasets>=2.14.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
nltk>=3.8.1